requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
pytest==7.1.1
requests==2.27.1
python-dotenv==0.20.0
loguru==0.6.0
SQLAlchemy==1.4.32
psycopg2==2.9.3 